requirements.txt
google-generativeai==0.5.4
langchain==0.2.1
langchain-google-genai==1.0.4
sentence-transformers==2.7.0

Flask==3.0.3
python-dotenv==1.0.1

requests==2.31.0
beautifulsoup4==4.12.3
PyPDF2==3.0.1
selenium==4.21.0
webdriver-manager==4.0.1

scikit-learn==1.4.2
numpy==1.26.4
nltk==3.8.1